Design updated, Fixed git conflicts
Design updated, Fixed git conflicts
