+++ .gitignore
@@ -16,7 +16,6 @@ wp-content/themes/twenty*/
 
 # ignore node dependency directories
 node_modules/
-public/node_modules/
 
 # ignore log files and databases
 *.log
