+++ package.json
@@ -1,29 +1,35 @@
 {
   "name": "addonify-compare-products",
   "version": "1.0.0",
-  "description": "Addonify WooCoomerce Quick View plugin adds functionality to have a WooCoomerce product quick view preview on a modal window.",
-  "main": "index.js",
+  "description": "Addonify WooCommerce Compare Products",
+  "main": "gulpfile.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "gulp"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/addonify/addonify-compare-products.git"
   },
-  "keywords": [
-    "woocommerce",
-    "quick",
-    "view",
-    "addonify"
-  ],
-  "author": "Addonify",
+  "author": "addonify",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/addonify/addonify-compare-products/issues"
   },
   "homepage": "https://github.com/addonify/addonify-compare-products#readme",
   "devDependencies": {
+    "autoprefixer": "^10.4.2",
+    "cssnano": "^5.1.1",
     "gulp": "^4.0.2",
-    "gulp-wp-pot": "^2.4.3"
+    "gulp-concat": "^2.6.1",
+    "gulp-notify": "^4.0.0",
+    "gulp-plumber": "^1.2.1",
+    "gulp-postcss": "^9.0.1",
+    "gulp-rename": "^2.0.0",
+    "gulp-replace": "^1.1.3",
+    "gulp-rtlcss": "^1.4.2",
+    "gulp-sass": "^5.1.0",
+    "gulp-sourcemaps": "^3.0.0",
+    "gulp-uglify": "^3.0.2",
+    "sass": "^1.49.9"
   }
 }
