+++ public/assets/src/scss/abstracts/_variables.scss
@@ -1,6 +1,5 @@
-
 // define few variables 
 
 $bck_color: rgb(41, 42, 48);
 $padding: 15px;
-$thumbnail_height : 60px;
+$thumbnail_height : 42px;
