--- public/assets/src/scss/addonify-compare-public.scss
@@ -1,338 +1,0 @@
-$bck_color: rgb(41, 42, 48);
-$padding: 15px;
-$thumbnail_height : 60px;
-
-#addonify-compare-modal-overlay,
-#addonify-compare-search-modal-overlay {
-    position: fixed;
-    left: 0;
-    right: 0;
-    top: 0;
-    bottom: 0;
-    background-color: rgba(0,0,0,0.8);
-    z-index: 9999;
-}
-
-.addonify-overlay-buttons{
-    position: relative;
-
-    .addonify-overlay-btn-wrapper{
-        position: absolute;
-        width: 100%;
-        top: 0;
-        left: 0;
-        background: #1010101a;
-        opacity: 0;
-        transition: 0.4s;
-
-        flex-direction: column;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-}
-
-
-button.addonify-cp-button.selected{
-    border: solid 2px #000;
-}
-
-#addonify-compare-footer{
-    position: fixed;
-    bottom: 0;
-    background: $bck_color;
-    padding: $padding;
-    right: 0;
-    transition: 0.4s;
-    z-index: 9999991;
-    display: none;
-}
-
-.addonify-compare-hidden{
-    display: none;
-}
-
-body.addonify-compare-footer-is-visible{
-    #addonify-compare-footer{
-        display: block;
-    }
-}
-
-#addonify-compare-footer-message{
-    float: left;
-    color: #9c9c9c;
-    padding-top: 12px;
-}
-
-#addonify-compare-footer-inner, #addonify-footer-thumbnails{
-    float: right;
-    overflow: hidden;
-    height: 100%;
-}
-
-#addonify-compare-footer-inner {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    justify-content: flex-end;
-    align-items: center;
-}
-
-#addonify-footer-thumbnails {
-    display: flex;
-    flex-wrap: nowrap;
-}
-
-.addonify-footer-components {
-    margin-left: 10px;
-
-    img{
-        height: 100%;
-        width: 100%;
-    }
-
-    button{
-        height: $thumbnail_height;
-    }
-}
-
-.addonify-footer-thumbnail{
-    position:relative;
-    cursor: move;
-    width: $thumbnail_height;
-    height: $thumbnail_height;
-    background: #3c3c3c;
-
-    &:hover{
-        .addonify-footer-remove.addonify-footer-btn{
-            display: block;
-        }
-    }
-}
-
-.addonify-footer-thumbnail.loading{
-    background-image: url(../../images/loading-dots.gif);
-    background-repeat: no-repeat;
-    background-position: center;
-}
-
-.addonify-footer-remove.addonify-footer-btn {
-    cursor: pointer;
-    top: 2px;
-    right: 2px;
-    width: 20px;
-    height: 20px;
-    position: absolute;
-    background-image: url(../../images/remove.svg);
-    background-repeat: no-repeat;
-    background-position: center;
-    background-color: rgba(0,0,0,0.7);
-    z-index: 7;
-}
-
-#addonify-footer-add{
-    width: $thumbnail_height;
-    height: $thumbnail_height;
-    display: block;
-    position: relative;
-    background-color: #3c3c3c;
-    background-image: url(../../images/add.svg);
-    background-size: 20px 20px;
-    background-repeat: no-repeat;
-    background-position: center;
-}
-
-.addonify-footer-settings{
-    width: $thumbnail_height;
-    height: $thumbnail_height;
-    display: block;
-    position: relative;
-    cursor: pointer;
-    background-color: transparent;
-    background-image: url(../../images/checkbox.svg);
-    background-size: 16px 16px;
-    background-repeat: no-repeat;
-    background-position: center;
-}
-
-
-// search modal
-#addonify-compare-search-modal{
-    position: fixed;
-    left: 50%;
-    top: 50%;
-    transform: translate(-50%, -50%);
-    z-index: 9999991;
-    border: 1px solid;
-}
-
-.addonify-compare-search-model-inner{
-    width: 360px;
-    height: 400px;
-    background: #fff;
-    position: relative;
-    padding-top: 25px;
-}
-
-.addonify-compare-search-head{
-    position: relative;
-}
-
-#addonify-compare-search-close-button{
-    position: absolute;
-    right: 0;
-    top: 0px;
-    padding: 5px;
-    line-height: 1;
-}
-
-.addonify-compare-search-modal-content{
-    padding: 15px;
-}
-
-#addonify-compare-search-query{
-    width: 100%;
-    border: solid 1px #000;
-    padding: 10px;
-    font-size: 16px;
-}
-
-
-#addonify-compare-search-results{
-    padding-top: 15px;
-    overflow-y: auto;
-    height: 328px;
-
-
-    &.loading{
-        background: url( ../../images/search.gif) no-repeat center;
-        min-height: 325px;
-    }
-
-    ul{
-        margin: 0;
-        padding: 0;
-    }
-
-    li{
-        list-style: none;
-        overflow: hidden;
-        border-bottom: solid 1px #eee;
-        padding-bottom: 10px;
-        margin-bottom: 10px;
-    }
-
-    li:last-child{
-        border-bottom: 0;
-        padding-bottom: 0;
-        margin-bottom: 0;
-    }
-
-    .item{
-        img.item-image{
-            width: 48px;
-            height: 48px;
-            float: left;
-            margin-right: 10px;
-        }
-
-        .item-name{
-            font-size: 18px;
-            padding-top: 11px;
-            float: left;
-            white-space: nowrap;
-            overflow: hidden;
-            text-overflow: ellipsis;
-        }
-
-        .item-add{
-            text-align: center;
-            float: right;
-            width: 32px;
-            height: 32px;
-            margin-top: 8px;
-            text-align: center;
-            color: #222;
-            background-color: #f2f2f2;
-            cursor: pointer;
-
-            span{
-                line-height: 31px;
-            }
-        }
-    }
-}
-
-
-
-// main display modal
-#addonify-compare-modal{
-    position: fixed;
-    left: 40px;
-    top: 40px;
-    right: 40px;
-    bottom: 40px;
-    background: #ffffff;
-    z-index: 999999;
-}
-
-.addonify-compare-model-inner{
-    position: relative;
-}
-
-#addonify-compare-close-button{
-    position: absolute;
-    right: 0;
-    top: 0;
-    z-index: 99;
-}
-
-
-
-// compare table contents
-#addonify-compare-products-table th,
-#addonify-compare-products-table td {
-    padding: 15px 15px;
-    max-width: 400px;
-    min-width: 200px;
-}
-
-
-#addonify-compare-modal-content {
-    height: calc(100vh - 80px);
-    overflow: auto;
-}
-
-.acp-field-title {
-    font-weight: bold;
-}
-
-.addonify-footer-remove {
-
-    margin-left: 15px;
-
-    button {
-        padding: 0px;
-        height: 30px;
-        width: 30px;
-        border-radius: 50%;
-    }
-}
-
-@media( max-width: 768px ) {
-    #addonify-compare-modal {
-        left: 15px;
-        right: 15px;
-        top: 15px;
-        bottom: 15px;
-    }
-
-    .addonify-footer-components button {
-        height: 45px;
-    }
-    .addonify-footer-thumbnail,
-    #addonify-footer-add {
-        height: 45px;
-        width: 45px;
-    }
-}
