+++ public/assets/src/scss/partials/components/_buttons.scss
@@ -1,4 +1,3 @@
-
 // button style
 
 .addonify-overlay-buttons {
@@ -12,14 +11,12 @@
     width: 100%;
     top: 0;
     left: 0;
-    background: #1010101a;
     opacity: 0;
-
     flex-direction: column;
     display: flex;
     align-items: center;
     justify-content: center;
-
+    background: #111111;
     @include transation_delay();
 }
