+++ public/templates/addonify-compare-products-compare-modal-wrapper.php
@@ -20,13 +20,11 @@
 // direct access is disabled.
 defined( 'ABSPATH' ) || exit;
 ?>
-
 <!-- product display modal -->
 <div id="addonify-compare-modal-overlay" class="addonify-compare-hidden"></div>
 
-<div id="addonify-compare-modal" class="addonify-compare-hidden" >
+<div id="addonify-compare-modal" class="addonify-compare-hidden">
 	<div class="addonify-compare-model-inner">
-
 		<button id="addonify-compare-close-button" class="addonify-compare-all-close-btn" >
 			<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
 				stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
@@ -34,8 +32,6 @@ defined( 'ABSPATH' ) || exit;
 				<line x1="6" y1="6" x2="18" y2="18"></line>
 			</svg>
 		</button>
-
-		<div id="addonify-compare-modal-content"></div>
-
+		<div id="addonify-compare-modal-content" class="adfy-magicscroll-bar"></div>
 	</div>
 </div>
